==> Bens_Code/01_Curate_Disney_Films_List.ipynb <==
# 01 Curate A List of Disney Films

I gathered Disney Film Lists from 
 - Disney.com (657 titles)
 - D23 (The official Disney Fan Club) (749 titles)
 - Wikipedia (425 titles)

The lists all contain different titles and different film data. I'll explore each and consider creating a single master list of Disney films to work from.

## 01.01 Python Imports

import requests
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import time

from pathlib import Path

from datetime import datetime
from sklearn.model_selection import train_test_split

import warnings # necessary b/c pandas & statsmodels datetime issue
warnings.simplefilter(action="ignore")

## 01.02 Disney.com Data

### 01.02.01 Import Disney.com data

FL_Disney = pd.read_csv('../Other Source Data/DD_Films_List_Disney_Com/Film_list_Disney_com.csv')
FL_Disney.rename(     columns=({ 'Movie Title': 'title'}),     inplace=True )

print(FL_Disney.shape)
FL_Disney.head()

### 01.02.02 Exploring Disney.com Data

In some cases, the remake of the film included the year of release in the title.  In some cases, the orignal release includes the year in the title. This may not be big deal as the IMDB

In some cases, parentheses do not indicate the year of release i.e Frozen (Sing-Along Edition) or The Wizards Return: Alex vs. Alex (TV Special)

FL_Disney[FL_Disney['title'].str.contains("\(")]

In some cases, the same or similar title is used multiple times. Since this list is from Disney.com, we will include all of these in our master list but may not include all of the films in our analyis.  Once we have more data, we will explore different groupings and determine which films or subset of films to work with.

FL_Disney[FL_Disney['title'].str.contains("101 Dalmatians")]

The 1982 release of Annie was produced by Rastart and distributed by Columbia pictures.  Not Disney. The 1999 version was a Disney production.

FL_Disney[FL_Disney['title'].str.contains("Annie")]

Cinderella was remade in 2015 as a live-action film. It seems odd that (1950) would be used to denote the original Disney production but we're going to leave it for now.

FL_Disney[FL_Disney['title'].str.contains("Cinderella")]

There was a Freaky Friday release in 1976, 2003, and 2018.  The 2018 release was a TV movie.  Not a theatrical release.

FL_Disney[FL_Disney['title'].str.contains("Freaky Friday")]

FL_Disney[FL_Disney['title'].str.contains("Pete's Dragon")]

FL_Disney[FL_Disney['title'].str.contains("Sleeping Beauty")]

When spot checking some title, I noticed that "Sing Along Songs: The Jungle Book -- The Bare ..." is in IMDB as "Disney Sing-Along-Songs: The Bare Necessities". Forthermore, when looking at IMDB, many of the "Disney Sing-Along-Songs" titles do not contain the name of the film. Creating a script to pick up the nuances may be time consuming with too litle reward. Sorting out the Sing-Along titles may be better done manually for the sake of time.

FL_Disney[FL_Disney['title'].str.contains("Jungle Book")]

FL_Disney[FL_Disney['title'].str.contains("Muppet Movie")]

FL_Disney[FL_Disney['title'].str.contains("Parent Trap")]

FL_Disney[FL_Disney['title'].str.contains("Shaggy Dog")]

FL_Disney[FL_Disney['title'].str.contains("Winnie the Pooh")]

## 01.03 D23 Data

### 01.03.01 Import D23 data

This one column holds the film title, year of release, and Motion Picture Association film rating. Will need to seperate those before moving forward.  <br>
Also, the D23 list (Disney's Official Fan Club) has almost 100 more titles that the Disney.com list.  657 to 749
.  I might need to create a list of titles that differ from list to list and research further

FL_D23 = pd.read_csv('../Other Source Data/DD_Films_List_D23/D23_list.csv', header=None )
FL_D23.rename(     columns=({ 0: 'orig_title'}),     inplace=True )
print(FL_D23.shape)
FL_D23.head(5)

### 01.03.02 Exploring D23 Data

# Splitting out the d_23 index number from the title
# Not dropping the d23_index just yet incase I need it to trouble shoot

index_split = FL_D23['orig_title'].str.split(".", n = 1, expand = True)    
FL_D23['d23_index'] = index_split[0]
FL_D23['d23_title_str'] = index_split[1]



FL_D23.head(5)

# Splitting out the year number from the title

year_split = FL_D23['d23_title_str'].str.split(":", n = 1, expand = True)    

FL_D23['d23_year'] = year_split[0]
FL_D23['d23_title_str'] = year_split[1]

FL_D23 = FL_D23[['orig_title', 'd23_index', 'd23_year', 'd23_title_str']]

FL_D23.head(5)

# Now that we have the year seperate, we have to split the title further.
# The problem is that some of the strings end with the rating (G), (PG-13), etc.
# While others have no rating
# Further complicating things, some of the title strings include the name of the production company in a set or parenthesis
# and lastly, sometimes there is a production company with a rating, sometimes there is a production company with no rating
# sometimes there is no production company, but there is a rating, and sometimes there is neither
# After a few different attempts, the logic just got too messy. Going to use if/else (np.where) statements

# If the rating is included in the title string, then assign the rating value to the rating column

FL_D23['d23_rating'] = np.where(FL_D23['d23_title_str'].str.contains("\(G\)"), "G", 
                                np.where(FL_D23['d23_title_str'].str.contains("\(PG\)"), "PG", 
                                         np.where(FL_D23['d23_title_str'].str.contains("\(PG-13\)"), "PG-13", 
                                                  np.where(FL_D23['d23_title_str'].str.contains("\(R\)"), "R",
                                                           np.where(FL_D23['d23_title_str'].str.contains("\(PG_13\)"), "PG-13",
                                                                    np.where(FL_D23['d23_title_str'].str.contains("\(PG-13"), "PG-13",
                                                                             np.where(FL_D23['d23_title_str'].str.contains("\(NR\)"), "NaN",
                                                           "NaN" ) ) ) ) ) ) )

# If the rating is included in the title string, then strip it out

FL_D23['d23_title_str'] = np.where(FL_D23['d23_title_str'].str.contains("\(G\)"), FL_D23['d23_title_str'].str.replace("\(G\)",''), 
                                np.where(FL_D23['d23_title_str'].str.contains("\(PG\)"), FL_D23['d23_title_str'].str.replace("\(PG\)",''), 
                                         np.where(FL_D23['d23_title_str'].str.contains("\(PG-13\)"), FL_D23['d23_title_str'].str.replace("\(PG-13\)",''), 
                                                  np.where(FL_D23['d23_title_str'].str.contains("\(R\)"), FL_D23['d23_title_str'].str.replace("\(R\)",''),
                                                           np.where(FL_D23['d23_title_str'].str.contains("\(PG_13\)"), FL_D23['d23_title_str'].str.replace("\(PG_13\)",''),
                                                                    np.where(FL_D23['d23_title_str'].str.contains("\(PG-13"), FL_D23['d23_title_str'].str.replace("\(PG-13",''),
                                                                             np.where(FL_D23['d23_title_str'].str.contains("\(NR\)"), FL_D23['d23_title_str'].str.replace("\(NR\)",''),
                                                                    FL_D23['d23_title_str'] ) ) ) ) ) ) )

FL_D23 = FL_D23[['orig_title', 'd23_index', 'd23_year', 'd23_rating', 'd23_title_str']]

FL_D23.head(35)

FL_D23.tail(35)

# Count the parenthesis remaining to understand how much clean up is left

FL_D23['d23_p_count'] = FL_D23['d23_title_str'].str.count("\(")

# 405 rows are ready to go
# 344 rows contain one set of parenthesis

FL_D23['d23_p_count'].value_counts()

FL_D23[FL_D23['d23_p_count'] == 1]

# Splitting out the title from everything else

studio_split = FL_D23['d23_title_str'].str.split("\(", n = 1, expand = True)

FL_D23['d23_title'] = studio_split[0]
FL_D23['d23_studio'] = studio_split[1]
FL_D23['d23_studio'] = FL_D23['d23_studio'].str.replace('\)', '')

# FL_D23[FL_D23['d23_p_count'] == 1]
FL_D23 = FL_D23[['orig_title', 'd23_index', 'd23_year', 'd23_rating', 'd23_title', 'd23_studio']]

FL_D23 = FL_D23[['orig_title', 'd23_index', 'd23_year', 'd23_rating', 'd23_title', 'd23_studio']]

FL_D23.head()

# spot checking some film titles to confirm that the splits all worked as expected.

# for val in FL_D23['d23_title']:
#     print(val)
    
# for val in FL_D23['d23_year']:
#     print(val)    

# There are a number of "Touchstone" and "Hollywood Pictures", so the split didn't work as expected
# for val in FL_D23['d23_rating']:
#     print(val)    

# for val in FL_D23['d23_studio']:
#     print(val)    

## 1.04 Wikipedia All Films

### 01.04.01 Import Wikipedia All Films  data

Wikipedia's list includes a lot more information, but only 425 titels, as compared to 657 from Disney.com and 748 from D23.  

FL_WK = pd.read_csv('../Other Source Data/DD_Films_LIst_WK/FIlms_List_WK.csv', )
print(FL_WK.shape)
FL_WK.head()

## 1.05 Wikipedia Animated Films

### 01.05.01 Import Wikipedia Animated  data

Wikipedia's list of Disney anitmated films.  A lot of movie info, but only 144 titles.

FL_WK_Animated = pd.read_csv('../Other Source Data/DD_Animated_List_WK/Disney_Animated_List_wk.csv', )
print(FL_WK_Animated.shape)
FL_WK_Animated.head()

## 1.06 Conclusion and next steps

For the sake of time, I elected to just work from the Disney.com list of films.  If Disney consideres a title to be a Disney movie, that that should be goog enough for me.
