==> Bens_Code/02.02_Connect_Other_IMDB_data.ipynb <==
import gzip
import pandas as pd

# nm=gzip.open('IMDB/name.basics.tsv.gz','rb')
# df_names = pd.read_csv(nm,sep='\t', low_memory=False)
# df_names.head(10)


# # Index nconst	primaryName	birthYear	deathYear	primaryProfession	knownForTitles
# # 0	nm0000001	Fred Astaire	1899	1987	soundtrack,actor,miscellaneous	tt0050419,tt0031983,tt0072308,tt0053137
# # 1	nm0000002	Lauren Bacall	1924	2014	actress,soundtrack	tt0117057,tt0071877,tt0038355,tt0037382
# # 2	nm0000003	Brigitte Bardot	1934	\N	actress,soundtrack,music_department	tt0057345,tt0049189,tt0056404,tt0054452
# # 3	nm0000004	John Belushi	1949	1982	actor,soundtrack,writer	tt0072562,tt0077975,tt0078723,tt0080455
# # 4	nm0000005	Ingmar Bergman	1918	2007	writer,director,actor	tt0083922,tt0069467,tt0060827,tt0050986
# # 5	nm0000006	Ingrid Bergman	1915	1982	actress,soundtrack,producer	tt0077711,tt0038109,tt0036855,tt0034583
# # 6	nm0000007	Humphrey Bogart	1899	1957	actor,soundtrack,producer	tt0037382,tt0043265,tt0042593,tt0034583
# # 7	nm0000008	Marlon Brando	1924	2004	actor,soundtrack,director	tt0078788,tt0068646,tt0070849,tt0047296
# # 8	nm0000009	Richard Burton	1925	1984	actor,soundtrack,producer	tt0057877,tt0087803,tt0061184,tt0059749
# # 9	nm0000010	James Cagney	1899	1986	actor,soundtrack,director	tt0031867,tt0029870,tt0042041,tt0035575

# aka=gzip.open('IMDB/title.akas.tsv.gz','rb')
# df_akas = pd.read_csv(aka,sep='\t', low_memory=False)
# df_akas.head(10)


# # index titleId	ordering	title	region	language	types	attributes	isOriginalTitle
# # 0	tt0000001	1	Карменсіта	UA	\N	imdbDisplay	\N	0
# # 1	tt0000001	2	Carmencita	DE	\N	\N	literal title	0
# # 2	tt0000001	3	Carmencita - spanyol tánc	HU	\N	imdbDisplay	\N	0
# # 3	tt0000001	4	Καρμενσίτα	GR	\N	imdbDisplay	\N	0
# # 4	tt0000001	5	Карменсита	RU	\N	imdbDisplay	\N	0
# # 5	tt0000001	6	Carmencita	US	\N	imdbDisplay	\N	0
# # 6	tt0000001	7	Carmencita	\N	\N	original	\N	1
# # 7	tt0000001	8	カルメンチータ	JP	ja	imdbDisplay	\N	0
# # 8	tt0000002	1	Le clown et ses chiens	\N	\N	original	\N	1
# # 9	tt0000002	2	Le clown et ses chiens	FR	\N	imdbDisplay	\N	0

# cr=gzip.open('IMDB/title.crew.tsv.gz','rb')
# df_crews = pd.read_csv(cr,sep='\t', low_memory=False)
# df_crews.tail(10)


# # index tconst	directors	writers
# # 8784762	tt9916838	nm5519375,nm5519454	nm6182221,nm1628284,nm2921377
# # 8784763	tt9916840	nm0996406	nm1482639,nm2586970
# # 8784764	tt9916842	nm5519454,nm5519375	nm6182221,nm1628284,nm2921377
# # 8784765	tt9916844	nm5519454,nm5519375	nm6182221,nm1628284,nm2921377
# # 8784766	tt9916846	nm5519375,nm5519454	nm6182221,nm1628284,nm2921377
# # 8784767	tt9916848	nm5519454,nm5519375	nm6182221,nm1628284,nm2921377
# # 8784768	tt9916850	nm5519454,nm5519375	nm6182221,nm1628284,nm2921377
# # 8784769	tt9916852	nm5519454,nm5519375	nm6182221,nm1628284,nm2921377
# # 8784770	tt9916856	nm10538645	nm6951431
# # 8784771	tt9916880	nm0996406	nm1482639,nm2586970

# since we are not considering anything other than full-length features, this may not be relevant
# ep=gzip.open('IMDB/title.episode.tsv.gz','rb')
# df_episodes = pd.read_csv(ep,sep='\t', low_memory=False)
# df_episodes.head()

# # Index tconst	parentTconst	seasonNumber	episodeNumber
# # 0	tt0020666	tt15180956	1	2
# # 1	tt0020829	tt15180956	1	1
# # 2	tt0021166	tt15180956	1	3
# # 3	tt0021612	tt15180956	2	2
# # 4	tt0021655	tt15180956	2	5

# pr=gzip.open('IMDB/title.principals.tsv.gz','rb')
# df_principals = pd.read_csv(pr,sep='\t', low_memory=False)
# df_principals.head()

# # Index tconst	ordering	nconst	category	job	characters
# # 0	tt0000001	1	nm1588970	self	\N	["Self"]
# # 1	tt0000001	2	nm0005690	director	\N	\N
# # 2	tt0000001	3	nm0374658	cinematographer	director of photography	\N
# # 3	tt0000002	1	nm0721526	director	\N	\N
# # 4	tt0000002	2	nm1335271	composer	\N	\N

rt=gzip.open('../Other Source Data/IMDB/title.ratings.tsv.gz','rb')
df_ratings = pd.read_csv(rt,sep='\t', low_memory=False)
df_ratings.head()
# Index	tconst	averageRating	numVotes
# 0	tt0000001	5.7	1868
# 1	tt0000002	5.9	247
# 2	tt0000003	6.5	1640
# 3	tt0000004	5.8	159
# 4	tt0000005	6.2	2463

Disney_IMDB = pd.read_csv('../Bens_Data/Combined_IMDB_Disney_453.csv')
Disney_IMDB.head()

Disney_IMDB.info()

Disney_IMDB = Disney_IMDB.rename(columns={'Unnamed: 0': 'orig_index','title': 'DFL_title'})

Disney_IMDB['endYear'].value_counts()

Disney_IMDB.drop(columns=['_merge','isAdult','endYear' ], inplace=True)

Pulling in the film rating from IMDB

Disney_IMDB = pd.merge(Disney_IMDB ,                 # left df
                          df_ratings,                  # right df
                          how="left",                 # left join
                          left_on='tconst',            # left column
                          right_on='tconst',    # right column
                          indicator = True,           # indicates source of each row
                          #validate = "one_to_many"    # alerts us of the relationship from left to right, incase there are dups
        )

Disney_IMDB['_merge'].value_counts()

11 a missing ratings.  Let's take a closer look at those.

Disney_IMDB[Disney_IMDB['_merge'] == 'left_only']

# This film has not yet been released (expected May 2022) tt9419884
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("doctor strange in the multiverse of madness")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt9419884"].index, inplace = True)

# Error made in the previous file.  tt15507514 should habe been dropped. 
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("holes")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt15507514"].index, inplace = True)

# tt11069282 This film has not yet been released (expected 2026) tt11069282 should habe been dropped. 
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("inspector gadget")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt11069282"].index, inplace = True)

# Error made in the previous file. This is not a Disney Movie.  The Disney film, tt2165933, was actually a TV movie. So neither of these should be in the list.
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("let it shine")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt7932304"].index, inplace = True)

# This film has not yet been released (expected June 2022) tt10298810
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("lightyear")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt10298810"].index, inplace = True)

# This film has not yet been released (expected 2024) tt11655566
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("lilo")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt11655566"].index, inplace = True)

# tt8292728 is not a Disney title and should be dropped
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("national")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt8292728"].index, inplace = True)

# tt1171257 is not a Disney movie
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("oceans")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt1171257"].index, inplace = True)

# tt1924347 was a TV movie, and should have been excluded
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("polly")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt1924347"].index, inplace = True)
Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['orig_index'] == "3787"].index, inplace = True)

# Error made in the previous file.  tt6476724 should habe been dropped.
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("the secret of the magic gourd")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt6476724"].index, inplace = True)

# This film has not yet been released (expected 2024) tt4872762
Disney_IMDB[Disney_IMDB['lower_title'].str.contains("the sword in the stone")]

Disney_IMDB.drop(Disney_IMDB[Disney_IMDB['tconst'] == "tt4872762"].index, inplace = True)

Disney_IMDB.info()

Disney_IMDB['titleType'].value_counts()

Disney_IMDB.to_csv('../Bens_Data/Updated_IMDB_Disney_442.csv')
