==> Bens_Code/06_Explore_Principal_Data.ipynb <==
# 06 Explore Principals Data


## 06.01 Imports


### 06.01.01 Python Imports


import pandas as pd
import numpy as np
import gzip

### 06.01.02 Import Principal History Data


History = pd.read_csv('../Bens_Data/imdb_scrape_director_writer_hist.csv')
History.drop(columns=['Unnamed: 0'], inplace=True)

History.info()

History.head()

## 06.02 Clean Up Principal History


### 06.02.01 Remove TV Shows


Since we are only concerned with movies, we want to drop anything that is TV or TV Related.  With more time, we would try to incporate some stats around TV history and then do some PCA to determine it would be worth including in our model.  Presumably, some history with TV might mean more success in film.

History['mpaarating'].value_counts()

History = History.drop(History[(History['mpaarating'].str.contains("TV"))].index)
History['mpaarating'].value_counts()

History.info()

### 06.02.02 Exploring MPAA Ratings


### 'Approved' and 'Passed' pre-date the current MPAA rating system.  
 - 'Passed' what was known as the 'Haynes Code' (1930-1934) 
 - 'Open' was also used in 1934
 - 'Approved' for exhibition (1934-1968)   
 - G, M, R, and X were used from 1968 to 1970 'M' meant "Suggested for mature audiences - Parental discretion advised"
 - G, GP, R, X were used from 1970 to 1972 'GP' meant "All ages admitted – Parental guidance suggested."
 - G, PG, R, X were used from 1973 to 1984 PG meant "Parental guidance suggested – Some material may not be suitable for pre-teenagers."
 - PG-13 was added in 1984 PG-13 meant "Parents strongly cautioned – Some material may be inappropriate for children under 13"
 - NC-17 Replaced X in 1990
 - M, M/PG were used by some films originating outside the U.S.
 
### We can probably smooth these out like this, or just create dummy columns.
  - 1 : Passed, Approved, Open, G
  - 2 : GP, PG
  - 2.5 : PG-13
  - 3 : R
  - 4 : NC-17, X
  - 100 : Not Rated, Unrated, and error (error meant the field was missing entirely from IMDB)
  - REMOVE : M, M/PG 
 
 With help from https://en.wikipedia.org/wiki/Motion_Picture_Association_film_rating_system#Replacement_of_the_Hays_Code

History[History['mpaarating'] == 'Not Rated']

History[History['mpaarating'] == 'Open']

History[History['mpaarating'] == 'M']

History[History['mpaarating'] == 'M/PG']

### 06.02.03 Exploring Release Dates


We do have release information for most films, not all.  

History[History['rlsdt'] == 'error']

### 06.02.04 Exploring Budget Data


We are missing a lot of budget information.  

History[History['budget'] == 'error']

### 06.02.05 Exploring Revenue Data


Of our 7765 rows, 5569 have no worldwide revenue numbers.

History[History['wordlwide'] == 'error']

## 06.03 IMDB Rating of Principal History


Given how much information is missing from our writer and director history, instead of trying to show a track record with ROI, maybe we can use just the revenue numbers and the IMDB rating of their previous work.

### 06.03.01 IMDB Rating Imports


rt=gzip.open('../Other Source Data/IMDB/title.ratings.tsv.gz','rb')
df_ratings = pd.read_csv(rt,sep='\t', low_memory=False)
df_ratings.head()
# # Index	tconst	averageRating	numVotes
# # 0	tt0000001	5.7	1868
# # 1	tt0000002	5.9	247
# # 2	tt0000003	6.5	1640
# # 3	tt0000004	5.8	159
# # 4	tt0000005	6.2	2463

### 06.03.02 Combining History & Rating Imports


History_Updated = pd.merge(History ,                 # left df
                          df_ratings,                  # right df
                          how="left",                 # left join
                          left_on='tconst',            # left column
                          right_on='tconst',    # right column
                          indicator = True,           # indicates source of each row
                          #validate = "one_to_many"    # alerts us of the relationship from left to right, incase there are dups
        )

History_Updated.info()

History_Updated['_merge'].value_counts(normalize=True)

### 06.03.03 Exploring Errors


History_Updated[(History_Updated['averageRating'].isnull() == 1) & (History_Updated['rlsdt'] != "error")].head(25)

Given that most of these are from the 1920's (long before IMDB) or they are in pre-development (not yet released), I'm ok still using this data set.  We can do some analysis on the which data is missing and maybe devise a method to collect that data in the future.

### 06.03.03 Removing TV data from Combined History


We previsouly removed 72 rows above for being TV projects. We'll need to remove those from our combined history table as well.

directors_writers_combined_history = pd.read_csv('../Bens_Data/directors_writers_combined_history.csv')
directors_writers_combined_history.drop(columns='Unnamed: 0',inplace=True)

directors_writers_combined_history.head()

## 06.04 Merge History List with Stats from Each Film


directors_writers_combined_history = pd.merge(directors_writers_combined_history ,                 # left df
                          History_Updated[['tconst', 'newurl', 'mpaarating', 'rlsdt', 'budget', 'wordlwide', 'averageRating', 'numVotes']],                  # right df
                          how="left",                 # left join
                          left_on='tconst',            # left column
                          right_on='tconst',    # right column
                          indicator = True,           # indicates source of each row
                          #validate = "one_to_many"    # alerts us of the relationship from left to right, incase there are dups
        ) 

### 06.04.01 Double Checknig Merge


directors_writers_combined_history.head()

directors_writers_combined_history.info()

directors_writers_combined_history._merge.value_counts(normalize=True)

directors_writers_combined_history = directors_writers_combined_history.drop(directors_writers_combined_history[(directors_writers_combined_history['_merge'].str.contains("left_only"))].index)

directors_writers_combined_history._merge.value_counts(normalize=True)

directors_writers_combined_history._merge.value_counts()

## 06.05 Export Directors/Writers Combined History Data


directors_writers_combined_history.to_csv('../Bens_Data/directors_writers_combined_history_updated.csv')
