==> Bens_Code/07_EDA_Clean_Principal_Data.ipynb <==
# 07 EDA and Clean Principals Data


## 07.01 Imports


### 07.01.01 Python Imports


import pandas as pd
import numpy as np
import re
import datetime

### 07.01.02 Import Principal History Data


principals_history = pd.read_csv('../Bens_Data/directors_writers_combined_history_updated.csv')
principals_history.drop(columns=['Unnamed: 0','_merge'], inplace=True)

principals_history.info()

To demonstrate or include a principal's trackrecord as part of our model, we're going to focus on the start year (so that get ratings for previous projects), genre (so we can get ratings for similar projects), average rating, and numvotes. We will attempt to use the rlsdat when we can but it might prove too dificult.  We also have too much missing data from the budget and worldwide columns to use those for now.  

## 07.02 Average Ratings


we'll remove any missing the averageRating and/or numVotes. We're not going to attempt to guess or impute the ratings or number of votes.  We're just going to remove those that are missing.

principals_history.head(10)

principals_history['averageRating'].isnull().sum(), principals_history['numVotes'].isnull().sum()

principals_history.drop(principals_history[principals_history['averageRating'].isnull()].index, inplace = True)

principals_history['averageRating'].isnull().sum(), principals_history['numVotes'].isnull().sum()

## 07.03 Start Year

principals_history['startYear'].isnull().sum(), principals_history['startYear'].str.startswith('\\N').sum()

## 07.04 Genres

We're not going to attempt to guess or impute the genres.  However, we would like to use them if at all possible. Any genre data that is missing will be filled with "unknown". We will do some vectorization on these columns later so that we can try to get track records for similar categories when building our final model.

principals_history['genres'].isnull().sum(), principals_history['genres'].str.startswith('\\N').sum()

principals_history['genres'].value_counts()

principals_history[principals_history['genres'].str.startswith('\\N')]

principals_history.loc[(principals_history['genres'].str.startswith('\\N')), 'genres'] = 'unknown'

principals_history['genres'].isnull().sum(), principals_history['genres'].str.startswith('\\N').sum()

## 07.05 Run Time

principals_history[principals_history['runtimeMinutes'].str.startswith('\\N')]

There remains 157 films without run times.  After spot checking a few, some were made, some were never released, some were released but not in the States.  Some border on pornography.   We may or may not use movies of similar runtimes, so we're just going to leave the data in here for now.


principals_history.info()

## 07.06 Budget and Worldwide Revenue


Budget and Worldwide should be numbers, not object.  Let's figure out what's going on.

principals_history['budget'].isnull().sum(), principals_history['budget'].str.startswith('\\N').sum(), principals_history['budget'].str.contains('error').sum()

principals_history['wordlwide'].isnull().sum(), principals_history['wordlwide'].str.startswith('\\N').sum(), principals_history['wordlwide'].str.contains('error').sum()

principals_history['wordlwide'].value_counts()

It looks like 'error' is the only thing preventing 'wordlwide' from being reconginzed as a number.

principals_history['wordlwide'].head(60)

principals_history['wordlwide'].tail(60)

Similarly, it looks like 'error' and the trailing '(estimated)' are preventing 'budget' from being reconginzed as a number.

principals_history['budget'].head(60)

principals_history['budget'].tail(60)

If we removed the (esitmated), and wrote the formula so that it skipped 'error', we could still get ROI for some 2,000 films.

principals_history.head()

principals_history['budget'].head(50)

### 07.06.01 Clean Up Budget Data

# Remove 'estimated'
principals_history['budget_adj'] = principals_history['budget'].apply(lambda x: x[:-12] if "estimated" in x else x)
principals_history['budget_adj'].head(50)

# Remove commas
principals_history['budget_adj'] = principals_history['budget_adj'].apply(lambda x: x.replace(",",""))
principals_history['budget_adj'] = principals_history['budget_adj'].apply(lambda x: x if "error" in x else x[1:])
principals_history['budget_adj'].head()

# Remove 'error'
principals_history['budget_adj'] = principals_history['budget_adj'].apply(lambda x: 0 if "error" in x else x)
principals_history['budget_adj'].head()

# Convert budget to integers
def convert(val):
    try:
        return(int(val))
    except:
        return(0)


principals_history['budget_adj'] = principals_history['budget_adj'].apply(lambda x: convert(x)) 

principals_history['budget_adj']

### 07.06.02 Clean Up Wordlwide Revenue Data

# Perform the same cleanup on Worldwide, remove commas, remove errors, then convert
principals_history['worldwide_adj'] = principals_history['wordlwide']
principals_history['worldwide_adj'] = principals_history['worldwide_adj'].apply(lambda x: x.replace(",",""))
principals_history['worldwide_adj'] = principals_history['worldwide_adj'].apply(lambda x: x if "error" in x else x[1:])
principals_history['worldwide_adj'] = principals_history['worldwide_adj'].apply(lambda x: 0 if "error" in x else x)

def convert(val):
    try:
        return(int(val))
    except:
        return(0)


principals_history['worldwide_adj'] = principals_history['worldwide_adj'].apply(lambda x: convert(x)) 
principals_history['worldwide_adj']

principals_history.head()

## 07.07 Create ROI Calculation


principals_history['ROI'] = (principals_history.worldwide_adj - principals_history.budget_adj ) / principals_history.budget_adj

principals_history.head()

## 07.08 Release Date


# Reformat Date
principals_history['rlsdt_dt'] = principals_history['rlsdt'].str.replace(r"\(.*\)","")

# Convert date format to get the date, month, and day information
principals_history['rlsdt_dt']= pd.to_datetime(principals_history['rlsdt_dt'],errors='coerce')
principals_history['rlsdt_mo']= pd.DatetimeIndex(principals_history['rlsdt_dt']).month
principals_history['rlsdt_day']= pd.DatetimeIndex(principals_history['rlsdt_dt']).day
principals_history['rlsdt_daynm']= pd.to_datetime(principals_history['rlsdt_dt']).dt.day_name()



principals_history.head(5)

principals_history.info()

Now that we have ROI and worldwide revenue for most films in the history, we need to summarize as best we can and get this data combined with Disney Feature Film data so that we can start building our models.    We could also get to runtime, month of release, day(date) and day of the week of release.  

## 07.09 Export Principal History


principals_history.to_csv('../Bens_Data/principals_history_updated.csv')
