==> Bens_Code/08_Update_Directors_Writers_List.ipynb <==
# 08 Update Directors & Writers Lists


## 08.01 Imports


### 08.01.01 Python Imports


import pandas as pd
import numpy as np

### 08.01.02 Directors & Writers Imports


directors_demo_history = pd.read_csv('../Bens_Data/directors_history.csv')
writers_demo_history = pd.read_csv('../Bens_Data/writers_history.csv')
directors_demo_history.drop(columns=['Unnamed: 0'], inplace=True)
writers_demo_history.drop(columns=['Unnamed: 0'], inplace=True)

### 08.01.03 Prinicpals History Import


principals_history = pd.read_csv('../Bens_Data/principals_history_updated.csv')
principals_history.drop(columns=['Unnamed: 0'], inplace=True)

directors_demo_history.head()

writers_demo_history.head()

principals_history.head()

## 08.02 Merging Data


### 08.02.01 Merging Directors and Principal History


directors_demo_history = pd.merge(directors_demo_history ,                 # left df
                          principals_history[['tconst','primaryTitle', 'runtimeMinutes', 'genres', 'mpaarating', 'rlsdt_dt', 'startYear','rlsdt_mo', 'rlsdt_day', 'rlsdt_daynm','budget_adj','worldwide_adj','ROI', 'averageRating', 'numVotes']],                  # right df
                          how="left",                 # left join
                          left_on='tconst',            # left column
                          right_on='tconst',    # right column
                          indicator = True,           # indicates source of each row
                          #validate = "one_to_many"    # alerts us of the relationship from left to right, incase there are dups
        ) 

### 08.02.02 Mergind Writers and Principal History


writers_demo_history = pd.merge(writers_demo_history ,                 # left df
                          principals_history[['tconst','primaryTitle', 'runtimeMinutes', 'genres', 'mpaarating', 'rlsdt_dt', 'startYear','rlsdt_mo', 'rlsdt_day', 'rlsdt_daynm','budget_adj','worldwide_adj','ROI', 'averageRating', 'numVotes']],                  # right df
                          how="left",                 # left join
                          left_on='tconst',            # left column
                          right_on='tconst',    # right column
                          indicator = True,           # indicates source of each row
                          #validate = "one_to_many"    # alerts us of the relationship from left to right, incase there are dups
        ) 

directors_demo_history.head()

writers_demo_history.head()

directors_demo_history['_merge'].value_counts()

writers_demo_history['_merge'].value_counts()

# For Directors, we probably should have done a left join as we only want the Directros data
directors_demo_history=directors_demo_history[directors_demo_history['_merge'] == 'both']

# ForWriters, we probably should have done a left join as we only want the Writers data
writers_demo_history=writers_demo_history[writers_demo_history['_merge'] == 'both']

directors_demo_history.info()

## 08.03 Clean Up Merged Data


### 08.03.01 Birth Year and Run Time


directors_demo_history['birthYear'] = pd.to_numeric(directors_demo_history['birthYear'],errors='coerce')
directors_demo_history['runtimeMinutes'] = pd.to_numeric(directors_demo_history['runtimeMinutes'],errors='coerce')
writers_demo_history['birthYear'] = pd.to_numeric(writers_demo_history['birthYear'],errors='coerce')
writers_demo_history['runtimeMinutes'] = pd.to_numeric(writers_demo_history['runtimeMinutes'],errors='coerce')

directors_history.info()

### 08.03.02 Calculate Age at Date of Each Project


directors_demo_history['age'] = directors_demo_history['startYear'] - directors_demo_history['birthYear']

writers_demo_history['age'] = writers_demo_history['startYear'] - writers_demo_history['birthYear']

directors_demo_history.info()

writers_demo_history.info()

The next step will be to add all of the demographic and history data, summarized, to out Disney movie data.

## 08.04 Export Directors and Writers History

directors_demo_history.to_csv('../Bens_Data/directors_demo_history.csv')
writers_demo_history.to_csv('../Bens_Data/writers_demo_history.csv')
